Agrego todo lo que estaba en routes2.js a routes.js y modifico routes.js para que pueda exportar correctamente, asi como tambien app.js
